test: split PromQL conflict hotspots and add golden matrix
