+++ compatibility/prometheus/query-corpus/fragments/004-functions.yml
@@ -1,0 +1,104 @@
+
+  # Functions.
+  - query: '{{.simpleTimeAggrOp}}_over_time(demo_memory_usage_bytes[{{.range}}])'
+    variant_args: ['simpleTimeAggrOp', 'range']
+  - query: 'quantile_over_time({{.quantile}}, demo_memory_usage_bytes[{{.range}}])'
+    variant_args: ['quantile', 'range']
+  - query: 'timestamp(demo_num_cpus)'
+  - query: 'timestamp(timestamp(demo_num_cpus))'
+  - query: '{{.simpleMathFunc}}(demo_memory_usage_bytes)'
+    variant_args: ['simpleMathFunc']
+  - query: '{{.simpleMathFunc}}(-demo_memory_usage_bytes)'
+    variant_args: ['simpleMathFunc']
+  - query: '{{.extrapolatedRateFunc}}(nonexistent_metric[5m])'
+    variant_args: ['extrapolatedRateFunc']
+  - query: '{{.extrapolatedRateFunc}}(demo_cpu_usage_seconds_total[{{.range}}])'
+    variant_args: ['extrapolatedRateFunc', 'range']
+  - query: 'deriv(demo_disk_usage_bytes[{{.range}}])'
+    variant_args: ['range']
+  - query: 'predict_linear(demo_disk_usage_bytes[{{.range}}], 600)'
+    variant_args: ['range']
+  - query: 'time()'
+    # label_replace does a full-string match and replace.
+  - query: 'label_replace(demo_num_cpus, "job", "destination-value-$1", "instance", "demo.promlabs.com:(.*)")'
+    # label_replace does not do a sub-string match.
+  - query: 'label_replace(demo_num_cpus, "job", "destination-value-$1", "instance", "host:(.*)")'
+    # label_replace works with multiple capture groups.
+  - query: 'label_replace(demo_num_cpus, "job", "$1-$2", "instance", "local(.*):(.*)")'
+    # label_replace resolves `$10` as capture group TEN, not group one followed
+    # by a literal `0`: Go's Regexp.ExpandString consumes the longest run of
+    # name characters after the `$`. ClickHouse's own regex substitution tops
+    # out at `\9`, so a backref above that ceiling has to be expanded by
+    # cerberus rather than handed to the engine.
+  - query: 'label_replace(demo_num_cpus, "job", "group10-$10", "instance", "(.)(.)(.)(.)(.)(.)(.)(.)(.)(.).*")'
+    # label_replace does not overwrite the destination label if the source label does not exist.
+  - query: 'label_replace(demo_num_cpus, "job", "value-$1", "nonexistent-src", "source-value-(.*)")'
+    # label_replace overwrites the destination label if the source label is empty, but matched.
+  - query: 'label_replace(demo_num_cpus, "job", "value-$1", "nonexistent-src", "(.*)")'
+    # label_replace does not overwrite the destination label if the source label is not matched.
+  - query: 'label_replace(demo_num_cpus, "job", "value-$1", "instance", "non-matching-regex")'
+    # label_replace drops labels that are set to empty values.
+  - query: 'label_replace(demo_num_cpus, "job", "", "dst", ".*")'
+    # label_replace fails when the regex is invalid.
+  - query: 'label_replace(demo_num_cpus, "job", "value-$1", "src", "(.*")'
+    should_fail: true
+    # `label_replace(..., "~invalid", ...)` / `label_replace(..., "instance",
+    # "", "", "")` / `label_join(..., "~invalid", ...)` were `should_fail`
+    # against Prom 2.x — `~invalid` violated the legacy label-name regex
+    # `[a-zA-Z_][a-zA-Z0-9_]*` so the function errored. Prom 3.0+ ships UTF-8
+    # label names by default
+    # (https://prometheus.io/docs/prometheus/latest/feature_flags/#utf-8-label-name-support),
+    # so `~invalid` is now legal and the queries succeed. Reference Prom in
+    # this harness is `prom/prometheus:v3.0.1`, so they can't be flagged
+    # `should_fail: true` anymore. Re-introduce only if the upstream
+    # compliance suite tightens these into Prom-3.x-shaped should_fail
+    # cases (e.g. genuinely malformed regex, multi-byte invalid identifier).
+  - query: 'label_join(demo_num_cpus, "new_label", "-", "instance", "job")'
+  - query: 'label_join(demo_num_cpus, "job", "-", "instance", "job")'
+  - query: 'label_join(demo_num_cpus, "job", "-", "instance")'
+  - query: '{{.dateFunc}}()'
+    variant_args: ['dateFunc']
+  - query: '{{.dateFunc}}(demo_batch_last_success_timestamp_seconds offset {{.offset}})'
+    variant_args: ['dateFunc', 'offset']
+  - query: '{{.instantRateFunc}}(demo_cpu_usage_seconds_total[{{.range}}])'
+    variant_args: ['instantRateFunc', 'range']
+  - query: '{{.clampFunc}}(demo_memory_usage_bytes, 2)'
+    variant_args: ['clampFunc']
+  - query: 'clamp(demo_memory_usage_bytes, 0, 1)'
+  - query: 'clamp(demo_memory_usage_bytes, 0, 1000000000000)'
+  - query: 'clamp(demo_memory_usage_bytes, 1000000000000, 0)'
+  - query: 'clamp(demo_memory_usage_bytes, 1000000000000, 1000000000000)'
+  - query: 'resets(demo_cpu_usage_seconds_total[{{.range}}])'
+    variant_args: ['range']
+  - query: 'changes(demo_batch_last_success_timestamp_seconds[{{.range}}])'
+    variant_args: ['range']
+    # A NaN-NaN run inside the lookback window: pins the changes() Prom
+    # parity carve-out for NaN-on-both-sides pairs (#1489) against a real
+    # reference Prometheus, not just chDB. See demo_gauge_with_nan_run in
+    # cmd/seed/main.go.
+  - query: 'changes(demo_gauge_with_nan_run[{{.range}}])'
+    variant_args: ['range']
+  - query: 'vector(1.23)'
+  - query: 'vector(time())'
+  - query: 'histogram_quantile({{.quantile}}, rate(demo_api_request_duration_seconds_bucket[1m]))'
+    variant_args: ['quantile']
+  - query: 'histogram_quantile(0.5, sum by (k8s_namespace_name, le) (rate(demo_resource_latency_seconds_bucket[1m])))'
+  - query: 'histogram_quantile(0.9, nonexistent_metric)'
+  - # Missing "le" label.
+    query: 'histogram_quantile(0.9, demo_memory_usage_bytes)'
+  - # Missing "le" label only in some series of the same grouping.
+    query: 'histogram_quantile(0.9, {__name__=~"demo_api_request_duration_seconds_.+"})'
+  - # #1483: the bare classic-histogram name (the OTel-CH *storage* name, no
+    # `_bucket` suffix) is not a Prom-wire series name — a classic histogram
+    # only exposes `<base>_bucket` / `_count` / `_sum`. Reference answers
+    # empty; this pins that cerberus does too instead of over-answering by
+    # scanning the same rows the `_bucket` spelling would.
+    query: 'histogram_quantile(0.9, demo_api_request_duration_seconds)'
+  - # #1692: `histogram_quantile(phi, sum by(le)(<bucket-selector>))` with no
+    # wrapping rate()/increase()/sum_over_time() is still valid PromQL — the
+    # instant vector's own sample values ARE the bucket counts. This must
+    # answer, not silently fold to empty.
+    query: 'histogram_quantile(0.9, sum by (le) (demo_api_request_duration_seconds_bucket))'
+  - query: 'count_values("value", demo_api_request_duration_seconds_bucket)'
+  - query: 'absent(demo_memory_usage_bytes)'
+  - query: 'absent(nonexistent_metric_name)'
