+++ compatibility/prometheus/query-corpus/manifest.txt
@@ -1,0 +1,8 @@
+000-scalars.yml
+001-vector-selectors.yml
+002-aggregation-operators.yml
+003-binary-and-unary-operators.yml
+004-functions.yml
+005-subqueries.yml
+006-native-histograms.yml
+007-duplicate-labelsets.yml
